Remove status components, include to list components #2 #5
