--- components/organisms/video/Status.vue
@@ -1,20 +1,0 @@
-<template>
-  <h3>
-    Video <span v-if="size">{{ size }} 件登録中</span>
-  </h3>
-</template>
-
-<script>
-export default {
-  props: {
-    size: {
-      type: Number,
-      required: true
-    }
-  }
-}
-</script>
-
-<style scoped>
-
-</style>
